ARM: ep93xx: add partial sparsemem support
Some boards, like for example ts-7300 with 128MB RAM as it seems can't be used
with flat memory model correctly. This patch adds possibility to choose
between flat/sparse memory models in kernel config and adds missing sparse
memory bits for ep93xx.

Patch is based on the previous discussion and code:

	http://marc.info/?l=linux-arm&m=122754446724900&w=2

I've tested this patch on the boards with 32MB and 128MB RAM.

[LOCAL, not for mainline!]

Signed-off-by: Petr Štetiar <ynezz@true.cz>