ARM: ep93xx: force nF bit on
ARM clocking mode is determined by the iA(31) and nF(30) bits, which are
usually set correctly by the bootloader or bootrom. If it is not set, then the
CPU core will run from HCLK instead of FCLK, and performance will suffer. It's
quite common case on some of the ts72xx boards.

[LOCAL, not for mainline!]

Signed-off-by: Petr Štetiar <ynezz@true.cz>