ARM: ep93xx: fix possibly wrong machine id passed from bootloader
In the early days Technologic Systems failed horribly and shipped some boards
(ts-72xx and ts-7800 are well known examples) with broken bootloaders, so
they've passed wrong machine id to the kernel.

We're trying to fix it with this patch in a way, that if we detect the wrong
machine id 0x163 we force it to correct machine id 0x2a1 (ts72xx).  Maybe,
that there are some other wrong numbers in the wild, so the code might need
some more tweaking.

[LOCAL, not for mainline!]

Signed-off-by: Petr Štetiar <ynezz@true.cz>