ARM: introduce atag_offset to replace boot_params
The boot_params member of the mdesc structure is used to provide a
default physical address for the ATAG list.  Since this value is fixed
at compile time and often based on ARCH_PHYS_OFFSET, it gets in the way
of runtime PHYS_OFFSET usage.

Let's introduce atag_offset which should contains only the relative
offset from start of memory instead of an absolute value, in preparation
to move all instance of boot_params over to it.

Signed-off-by: Nicolas Pitre <nicolas.pitre@linaro.org>