st7735fb: Working WIP changes to make DMA safe and add endian fix
This removes the "from the stack" allocation of 1 byte buffers
that was a temporary thing and isn't dma safe. Now vmallocs the
user buffer to be deferred I/O safe and kmallocs a swapped
shadow buffer to support the byte swabbing hack to fix userspace
limitations.

The buffer allocation and endian hack code needs some serious
cleanup as on big endian systems if would fail miserably atm.
However, the LE path works for the moment so people can do something
with the driver.

Signed-off-by: Matt Porter <mporter@ti.com>
Signed-off-by: Petr Štetiar <ynezz@true.cz>