ARM: dts: imx6: Fix PCIe reset gpio polarity on Toradex Apalis Ixora
Adding reset-gpio-active-high boolean DT binding property, which we need
to make PCIe working on Apalis SoMs and not break old DTBs. While at it,
I've fixed comment and GPIO polarity.

On Apalis SoMs the GPIO1_IO28 used to PCIe reset is not connected
directly to PERST# PCIe signal, but it's ORed with RESETBMCU coming off
the PMIC, and thus is inverted, active-high.

Signed-off-by: Petr Štetiar <ynezz@true.cz>